+++ src/app/app.module.ts
@@ -2,6 +2,7 @@ import { AppService } from './app.service';
 import { AppController } from './app.controller';
 import { Module, HttpModule } from '@nestjs/common';
 import { ModelGeneratorModule } from '../model-generator/model-generator.module';
+import { CreateGeneratorModule } from '../create-generator/create-generator.module';
 import { ModuleGeneratorModule } from '../module-generator/module-generator.module';
 import { ServiceGeneratorModule } from '../service-generator/service-generator.module';
 import { ControllerGeneratorModule } from '../controller-generator/controller-generator.module';
@@ -12,6 +13,7 @@ import { StaticFileGeneratorModule } from '../static-files-generator/static-file
   imports: [
     HttpModule,
     ModelGeneratorModule,
+    CreateGeneratorModule,
     ModuleGeneratorModule,
     ServiceGeneratorModule,
     RepositoryGeneratorModule,