+++ src/create-generator/create-generator.service.ts
@@ -1,0 +1,11 @@
+import { Injectable } from '@nestjs/common';
+import { DEFAULT_TAB_SIZE } from '../shared/constants';
+import { generateClass } from '../templates/create.template';
+import { CreateCreateDto } from '../shared/models/create-create.dto';
+
+@Injectable()
+export class CreateGeneratorService {
+  generateCreateDto(createCreateDto: CreateCreateDto) {
+    return generateClass(createCreateDto, DEFAULT_TAB_SIZE);
+  }
+}